+++ src/main/kotlin/blackjack/model/deck/Card.kt
@@ -1,6 +1,6 @@
 package blackjack.model.deck
 
-import blackjack.model.util.CardNumber
-import blackjack.model.util.Pattern
+import blackjack.util.CardNumber
+import blackjack.util.Pattern
 
 data class Card(val cardNumber: CardNumber, val pattern: Pattern)
