+++ src/main/kotlin/blackjack/model/deck/Deck.kt
@@ -1,7 +1,7 @@
 package blackjack.model.deck
 
-import blackjack.model.util.CardNumber
-import blackjack.model.util.Pattern
+import blackjack.util.CardNumber
+import blackjack.util.Pattern
 
 class Deck {
     private val cards: MutableList<Card>
